fix: resolve all 59 clippy warnings for clean lint baseline
Mechanical fixes across 22 files to pass cargo clippy -- -D warnings:
- map_or(false, ...) -> is_some_and(...)
- Collapsed nested if/else-if blocks
- Removed needless returns, unnecessary casts, excessive float precision
- Added #[allow(clippy::too_many_arguments)] on hot-path functions
- Replaced manual Default impls with #[derive(Default)]
- Used if-let instead of single-arm match, ? operator, div_ceil, etc.

+++ src/mesh_import.rs
@@ -54,7 +54,7 @@ pub fn load_obj(path: &Path) -> Result<TriMesh, String> {
             // Triangulate polygon (fan from first vertex)
             if indices.len() >= 3 {
                 for i in 1..indices.len() - 1 {
-                    triangles.push([indices[0], indices[i as usize], indices[i as usize + 1]]);
+                    triangles.push([indices[0], indices[i], indices[i + 1]]);
                 }
             }
         }
